+++ Source/RTEngine/Plane.hpp
@@ -9,16 +9,16 @@ namespace RealRT
     class Plane : public Shape
     {
         public:
-            Plane(std::shared_ptr<const Material> mat, const Vector3D &normal = {0.f, 0.f, 1.f}, float d = 5.f);
+            Plane(std::shared_ptr<const Material> mat, const Vector3D &normal = {0.0, 0.0, 1.0}, double d = 5.0);
 			~Plane(void) = default;
 
-		    float Intersect(const Ray &incident, bool &flipNormals) const;
+		    double Intersect(const Ray &incident, bool &flipNormals) const;
 		    Vector3D Normal(const Vector3D &pt) const;
 
 		    bool IsLight(void) const;
 
 		private:
             Vector3D _Normal;
-		    float _D;
+		    double _D;
     };
 }
