+++ Source/RTEngine/SphericalLight.cpp
@@ -2,7 +2,7 @@
 
 using namespace RealRT;
 
-SphericalLight::SphericalLight(std::shared_ptr<const Material> mat, const Vector3D &center, float radius)
+SphericalLight::SphericalLight(std::shared_ptr<const Material> mat, const Vector3D &center, double radius)
     : Sphere(mat, center, radius)
 {
