+++ Source/RTEngine/SphericalLight.hpp
@@ -8,7 +8,7 @@ namespace RealRT
     class SphericalLight : public Sphere
     {
     public:
-        SphericalLight(std::shared_ptr<const Material> mat, const Vector3D &center = {0.f, 0.f, 0.f}, float radius = 2.5f);
+        SphericalLight(std::shared_ptr<const Material> mat, const Vector3D &center = {0.0, 0.0, 0.0}, double radius = 2.5);
 
         bool IsLight(void) const;
     };
